split trips code into submodule, replace getCurrentTrip with PG variant
