+++ py/sql.py
@@ -62,7 +62,6 @@ adminStats = open("sql/stats/adminStats.sql", "r").read()
 leaderboardStats = open("sql/stats/leaderboardStats.sql", "r").read()
 typeAvailable = open("sql/stats/typeAvailable.sql", "r").read()
 getManualStationsQuery = open("sql/getManualStations.sql", "r").read()
-getCurrentTrip = open("sql/getCurrentTrip.sql", "r").read()
 getAirports = open("sql/getAirports.sql", "r").read()
 getTrainStations = open("sql/getTrainStations.sql", "r").read()
 getDuplicate = open("sql/getDuplicate.sql", "r").read()
