+++ src/trips/__init__.py
@@ -1,0 +1,26 @@
+from .create_trip import create_trip
+from .delete_trip import delete_trip
+from .duplicate_trip import duplicate_trip
+from .edits import (
+    attach_ticket_to_trips,
+    change_trips_visibility,
+    delete_ticket_from_db,
+    update_trip_type,
+)
+from .trip import Trip
+from .update_trip import update_trip
+from .utils import compare_trip, get_current_trip_id
+
+__all__ = [
+    Trip.__name__,
+    create_trip.__name__,
+    delete_trip.__name__,
+    duplicate_trip.__name__,
+    attach_ticket_to_trips.__name__,
+    change_trips_visibility.__name__,
+    update_trip.__name__,
+    delete_ticket_from_db.__name__,
+    update_trip_type.__name__,
+    compare_trip.__name__,
+    get_current_trip_id.__name__,
+]
