+++ templates/map.html
@@ -289,7 +289,7 @@ function integrateChanges(existingData, changes, currentTripIds) {
 var username = "{{username}}";
 localforage.getItem("lastLocal_" + username, function(error, lastLocal){
   if (lastLocal === null) {
-    $.get('{{ url_for("public_getTripsPaths", username=username, lastLocal="all") if public else url_for("getTripsPaths", username=username, lastLocal="all") }}', function(data, status){
+    $.get('{{ url_for("public_getTripsPaths", username=username, lastLocal="all") if public else url_for("get_trip_paths", username=username, lastLocal="all") }}', function(data, status){
       let trips = data.trips;
       placePolylines(trips);
       localforage.setItem("trips_" + username, trips);
@@ -300,7 +300,7 @@ localforage.getItem("lastLocal_" + username, function(error, lastLocal){
       displayPolylines();
     });
   } else {
-    $.get('{{ url_for("public_getTripsPaths", username=username, lastLocal="") if public else url_for("getTripsPaths", username=username, lastLocal="") }}'+lastLocal, function(data, status){
+    $.get('{{ url_for("public_getTripsPaths", username=username, lastLocal="") if public else url_for("get_trip_paths", username=username, lastLocal="") }}'+lastLocal, function(data, status){
       localforage.getItem("trips_" + username, function(error, storedTrips){
         let currentTripIds = data.idList;
         let updatedTrips = integrateChanges(storedTrips, data.trips, currentTripIds);
